// packages/plugin-assets-ui/src/common/types.ts
export type IAssetCategoryTypes = {
  _id: string;
  name: string;
  order: string;
  code: string;
  parentId: string;
  description: string;
  status: string;
  attachment: IAttachment;
  isRoot: boolean;
  assetCount: number;
};

export type IAssetCategoryQeuryResponse = {
  assetCategories: IAssetCategoryTypes[];
  loading: boolean;
  refetch: () => void;
};

export type IAssetCategoriesTotalCountResponse = {
  assetCategoriesTotalCount: number;
} & QueryResponse;

export type IAssetQueryResponse = {
  assets: IAsset[];
  loading: boolean;
  refetch: (variables?: any) => void;
};

export type IMovementDetailQueryResponse = {
  assetMovement: IMovementType;
  loading: boolean;
  refetch: () => void;
};

export type IAssetTotalCountQueryResponse = {
  assetsTotalCount: number;
  loading: boolean;
  refetch: () => void;
};
export type IAssetDetailQueryResponse = {
  assetDetail: IAsset;
  loading: boolean;
  refetch: () => void;
};
export interface IAssetCategory {
  _id: string;
  name: string;
  order: string;
  code: string;
  description?: string;
  attachment?: any;
  status: string;
  parentId?: string;
  createdAt: Date;
  assetCount: number;
  isRoot: boolean;
}

export interface IAsset {
  _id: string;
  name: string;
  type: string;
  categoryId: string;
  parentId: string;
  description: string;
  order: string;
  code: string;
  unitPrice: number;
  customFieldsData?: any;
  createdAt: Date;
  vendorId?: string;

  attachment?: any;
  attachmentMore?: any[];
  assetCount: number;
  minimiumCount: number;
  category: IAssetCategory;
  parent: IAsset;
  childAssetCount: number;
  vendor?: ICompany;
  kbArticleIds: string[];
  knowledgeData?: any;
}